requirements.txt
google-adk>=0.1.0
google-generativeai>=0.3.0
fastapi>=0.95.0
uvicorn>=0.22.0
python-dotenv>=1.0.0